XMLHttpRequest/event-load.htm

<!doctype html>
<meta charset=utf-8>
<title>XMLHttpRequest: load event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#handler-xhr-onload" data-tested-assertations="/../.." />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#event-xhr-load" data-tested-assertations="/../.." />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#switch-done" data-tested-assertations="/following-sibling::ol//li[6]" />
<div id="log"></div>

<script>
  var test = async_test();
  test.step(function() {
    var client = new XMLHttpRequest();
    client.onload = test.step_func(function() {
      assert_equals(client.readyState, 4);
      test.done();
    });
    client.onreadystatechange = test.step_func(function() {
      if (client.readyState !== 4) return;

      setTimeout(test.step_func(function() {
          assert_unreached("Didn't get load event within 4ms of readystatechange==4");
      }), 4);
    });
    client.open("GET", "resources/well-formed.xml");
    client.send(null);
  });
</script>
